Prevent deleting logged in user

+++ server/api/org-people/remove-user-from-org.delete.ts
@@ -3,9 +3,10 @@ import { organizations, organizationsPeople, users } from "../../db/schema"
 
 export default defineEventHandler(async (event) => {
     const db = event.context.drizzle
+    const currentUserId = event.context.auth.userId
     const { email, orgName } = await readBody(event)
 
-    if (await roleEditNotAllowed(event, orgName)) {
+    if (await roleEditNotAllowed(event, orgName)) {        
         throw createError({
             message: 'Unauthorized',
             statusCode: 401,
@@ -22,6 +23,13 @@ export default defineEventHandler(async (event) => {
     }).from(organizations)
         .where(eq(organizations.name, orgName))
         .then(takeUniqueOrThrow)
+    if (currentUserId === userIdFromEmail.userId) {
+        throw createError({
+            message: 'Unauthorized change your own role',
+            statusCode: 400,
+        })
+    }
+
     await db.delete(organizationsPeople)
         .where(and(
             eq(organizationsPeople.organizationId, orgIdFromOrgName.orgId),
